requirements.txt
beautifulsoup4==4.6.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-modelcluster==5.0
django-recaptcha==2.0.5
django-taggit==1.1.0
django-treebeard==4.3
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.8
Pillow==6.2.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.7
wagtail==2.7
wagtail-django-recaptcha==1.0
wagtailgmaps==1.0.1
webencodings==0.5.1
Willow==1.3
